Update the services and models

+++ trackzilla/src/main/java/com/luisenricke/trackzilla/entity/Application.java
@@ -8,7 +8,7 @@ public class Application {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "application_id")
-    private Integer id;
+    private Long id;
 
     @Column(name = "app_name", nullable = false)
     private String name;
@@ -26,18 +26,18 @@ public class Application {
         this.owner = owner;
     }
 
-    public Application(Integer id, String name, String description, String owner) {
+    public Application(Long id, String name, String description, String owner) {
         this.id = id;
         this.name = name;
         this.description = description;
         this.owner = owner;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
